fix: address approval payload review comments
