+++ internal/teammcp/actions_test.go
@@ -314,6 +314,32 @@ func TestBuildActionApprovalSpecPayloadEdgeCases(t *testing.T) {
 		}
 	})
 
+	t.Run("redacts secret-shaped keys inside array objects", func(t *testing.T) {
+		spec := buildActionApprovalSpec("growthops", "general", TeamActionExecuteArgs{
+			Platform: "gmail",
+			ActionID: "GMAIL_SEND_EMAIL",
+			Data: map[string]any{
+				"recipients": []any{
+					map[string]any{
+						"email":        "alex@nex.ai",
+						"access_token": "super-secret-token",
+						"headers": map[string]any{
+							"authorization": "Bearer nope",
+						},
+					},
+				},
+			},
+		})
+		for _, leaked := range []string{"super-secret-token", "Bearer nope", "access_token", "authorization", "headers"} {
+			if strings.Contains(spec.Context, leaked) {
+				t.Fatalf("approval context leaks redacted array value %q:\n%s", leaked, spec.Context)
+			}
+		}
+		if !strings.Contains(spec.Context, "alex@nex.ai") {
+			t.Fatalf("approval context dropped safe array value:\n%s", spec.Context)
+		}
+	})
+
 	t.Run("clips long bodies", func(t *testing.T) {
 		long := strings.Repeat("x", 800)
 		spec := buildActionApprovalSpec("growthops", "general", TeamActionExecuteArgs{
@@ -393,6 +419,24 @@ func TestBuildActionApprovalSpecPayloadEdgeCases(t *testing.T) {
 		}
 		if !strings.Contains(spec.Context, "primary@x.com") {
 			t.Fatalf("expected primary recipient, got:\n%s", spec.Context)
+		}
+	})
+
+	t.Run("direct field wins over nested envelope", func(t *testing.T) {
+		spec := buildActionApprovalSpec("growthops", "general", TeamActionExecuteArgs{
+			Platform: "gmail",
+			ActionID: "GMAIL_SEND_EMAIL",
+			Data: map[string]any{
+				"request": map[string]any{
+					"body": map[string]any{
+						"to": "nested@example.com",
+					},
+				},
+				"to": "top-level@example.com",
+			},
+		})
+		if !strings.Contains(spec.Context, "• To: top-level@example.com") {
+			t.Fatalf("expected direct recipient to win, got:\n%s", spec.Context)
 		}
 	})
